Eliminated more raw pointers in DMD classes. Added more const correctness.

+++ unit_tests/test_HDFDatabase.cpp
@@ -823,7 +823,7 @@ TEST(BasisGeneratorIO, Scaling_test)
     auto stop = steady_clock::now();
     auto duration = duration_cast<milliseconds>(stop-start);
     if (rank == 0)
-        printf("Base writeSnapshot- duration: %dms\n", duration.count());
+        printf("Base writeSnapshot- duration: %dms\n", (int) duration.count());
 
     MPI_Barrier(MPI_COMM_WORLD);
     start = steady_clock::now();
@@ -832,7 +832,7 @@ TEST(BasisGeneratorIO, Scaling_test)
     stop = steady_clock::now();
     duration = duration_cast<milliseconds>(stop-start);
     if (rank == 0)
-        printf("MPIO writeSnapshot- duration: %dms\n", duration.count());
+        printf("MPIO writeSnapshot- duration: %dms\n", (int) duration.count());
 
     MPI_Barrier(MPI_COMM_WORLD);
     start = steady_clock::now();
@@ -841,7 +841,7 @@ TEST(BasisGeneratorIO, Scaling_test)
     stop = steady_clock::now();
     duration = duration_cast<milliseconds>(stop-start);
     if (rank == 0)
-        printf("Base endSamples- duration: %dms\n", duration.count());
+        printf("Base endSamples- duration: %dms\n", (int) duration.count());
 
     MPI_Barrier(MPI_COMM_WORLD);
     start = steady_clock::now();
@@ -850,7 +850,7 @@ TEST(BasisGeneratorIO, Scaling_test)
     stop = steady_clock::now();
     duration = duration_cast<milliseconds>(stop-start);
     if (rank == 0)
-        printf("MPIO endSamples- duration: %dms\n", duration.count());
+        printf("MPIO endSamples- duration: %dms\n", (int) duration.count());
 }
 
 int main(int argc, char* argv[])
